server/requirements.txt
websockets==11.0.3
pynput==1.7.6
pywebview==4.2.2 # install following reqs in linux
# sudo apt install python3-gi python3-gi-cairo gir1.2-gtk-3.0 gir1.2-webkit2-4.0